pyproject.toml
[tool.poetry]
name = "mlops-city-concierge"
version = "0.1.0"
package-mode = false
description = "City Concierge — RAG application for city information (USF MSDS 603 MLOps)"
authors = ["James <pjnhek@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"

# Web framework
fastapi = ">=0.111.0,<1.0.0"
uvicorn = {version = ">=0.29.0,<1.0.0", extras = ["standard"]}

# Database / vector store
psycopg2-binary = ">=2.9.9,<3.0.0"
pgvector = ">=0.3.0,<1.0.0"
sqlalchemy = ">=2.0.29,<3.0.0"
alembic = ">=1.13.1,<2.0.0"

# AI / embeddings
openai = ">=1.30.0,<2.0.0"
google-genai = ">=0.7.0,<1.0.0"
langchain = ">=0.2.0,<1.0.0"
langchain-community = ">=0.2.0,<1.0.0"
langchain-google-genai = ">=1.0.0,<3.0.0"
langchain-openai = ">=0.1.0,<1.0.0"
mlflow = ">=2.12.0,<3.0.0"
tiktoken = ">=0.7.0,<1.0.0"

# Utilities
pydantic = ">=2.7.0,<3.0.0"
pydantic-settings = ">=2.2.1,<3.0.0"
python-dotenv = ">=1.0.1,<2.0.0"
httpx = ">=0.27.0,<1.0.0"
tenacity = ">=8.3.0,<9.0.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.2.0,<9.0.0"
pytest-asyncio = ">=0.23.6,<1.0.0"
pytest-cov = ">=5.0.0,<6.0.0"
pytest-mock = ">=3.14.0,<4.0.0"
factory-boy = ">=3.3.0,<4.0.0"
ruff = ">=0.4.4,<1.0.0"
mypy = ">=1.10.0,<2.0.0"
types-requests = ">=2.31.0,<3.0.0"
sqlalchemy = {version = ">=2.0.29,<3.0.0", extras = ["mypy"]}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --tb=short"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = false
ignore_missing_imports = true
